refactor: move SAFE_ACCOUNTS_LIMIT to spaces root, share at-limit notice copy
