+++ apps/web/src/features/spaces/components/AddAccounts/__tests__/AddAccounts.test.tsx
@@ -23,7 +23,8 @@ const setListOverflow = (overflow: boolean) => {
   })
 }
 
-jest.mock('../../Sidebar/constants', () => ({
+jest.mock('@/features/spaces/constants', () => ({
+  ...jest.requireActual('@/features/spaces/constants'),
   SAFE_ACCOUNTS_LIMIT: 10,
 }))
