+++ apps/web/src/features/spaces/components/AddAccountsChooser/index.tsx
@@ -7,8 +7,7 @@ import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/u
 import { cn } from '@/utils/cn'
 import AccountsModal from '@/components/common/SpaceSafeBar/AccountsModal'
 import AddAccounts from '@/features/spaces/components/AddAccounts'
-import { useCurrentSpaceId, useIsAdmin, useIsCurrentSpaceAtSafeLimit } from '@/features/spaces'
-import { SAFE_ACCOUNTS_LIMIT } from '@/features/spaces/components/Sidebar/constants'
+import { SAFE_ACCOUNTS_LIMIT, useCurrentSpaceId, useIsAdmin, useIsCurrentSpaceAtSafeLimit } from '@/features/spaces'
 import { trackEvent } from '@/services/analytics'
 import { SPACE_EVENTS } from '@/services/analytics/events/spaces'
 import { OVERVIEW_LABELS } from '@/services/analytics/events/overview'
