+++ apps/web/src/features/spaces/components/Sidebar/variants/SpaceSelectorDropdown/__tests__/SpaceSelectorDropdown.test.tsx
@@ -4,7 +4,8 @@ import type { ReactElement, ReactNode, CSSProperties } from 'react'
 import { AppRoutes } from '@/config/routes'
 import { trackEvent } from '@/services/analytics'
 import { getDeterministicColor } from '@/utils/colors'
-import { SPACE_SELECTOR_NAME_MAX_LENGTH, SPACES_LIMIT } from '../../../constants'
+import { SPACE_SELECTOR_NAME_MAX_LENGTH } from '../../../constants'
+import { SPACES_LIMIT } from '@/features/spaces/constants'
 import { truncateSpaceName } from '../../../utils'
 import { SpaceSelectorDropdown } from '../SpaceSelectorDropdown'
