+++ apps/web/src/features/spaces/components/SpacesList/index.tsx
@@ -31,7 +31,7 @@ import { AppRoutes } from '@/config/routes'
 import NextLink from 'next/link'
 import { useSignInRedirect } from '@/components/welcome/WelcomeLogin/hooks/useSignInRedirect'
 import AddIcon from '@/public/images/common/add.svg'
-import { SPACES_LIMIT } from '../Sidebar/constants'
+import { SPACES_LIMIT } from '@/features/spaces/constants'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip'
 import SafeLogo from '@/public/images/logo-no-text.svg'
 import { Pulse } from '../Sidebar/SidebarSkeleton/SidebarSkeleton'
