+++ apps/web/src/features/spaces/hooks/__tests__/useSelectAll.test.tsx
@@ -4,7 +4,8 @@ import { FormProvider, useForm, useFormContext } from 'react-hook-form'
 import type { AllSafeItems, MultiChainSafeItem, SafeItem } from '@/hooks/safes'
 import { MULTICHAIN_SAFE_KEY_PREFIX } from '../../components/SelectSafesOnboarding/constants'
 
-jest.mock('../../components/Sidebar/constants', () => ({
+jest.mock('../../constants', () => ({
+  ...jest.requireActual('../../constants'),
   SAFE_ACCOUNTS_LIMIT: 3,
 }))
